+++ src/main/java/br/com/postech/sevenfoodproduction/sevenfoodproductionapi/model/service/ProductionOrderService.java
@@ -3,7 +3,7 @@ package br.com.postech.sevenfoodproduction.sevenfoodproductionapi.model.service;
 import br.com.postech.sevenfoodproduction.sevenfoodproductionapi.model.entities.ProductionOrder;
 
 public interface ProductionOrderService {
-    ProductionOrder updateStatus(String orderId, Long status);
+    ProductionOrder updateStatus(String orderId, Integer status);
 
     Object findByOrderNumber(String orderId);
